nplot using base graphics

nplot using base graphics

nplot_base(x, layout = igraph::layout_nicely(x),
  vertex.size = igraph::degree(x, mode = "in"), bg.col = "lightgray",
  vertex.nsides = 50, vertex.color = NULL, vertex.size.range = c(0.01,
  0.03), vertex.frame.color = NULL, vertex.rot = 0,
  vertex.frame.prop = 0.1, edge.width = NULL, edge.width.range = c(1, 2),
  edge.arrow.size = NULL, edge.color = NULL, edge.color.mix = 0.5,
  edge.color.alpha = c(0.1, 0.5), edge.curvature = pi/3,
  edge.line.lty = "solid", edge.line.breaks = 15, sample.edges = 1,
  skip.vertex = FALSE, skip.edges = FALSE, skip.arrows = skip.edges,
  add = FALSE, zero.margins = TRUE)

Arguments

x
An igraph object.

layout
Numeric two-column matrix with the graph layout.

vertex.size
Numeric vector of length vcount(x). Absolute size of the vertex.

bg.col
Color of the background.

vertex.nsides
Numeric vector of length vcount(x). Number of sizes of the vertex. E.g. three is a triangle, and 100 approximates a circle.

vertex.color
Vector of length vcount(x). Vertex colors.

vertex.size.range
Vector of length vcount(x).

vertex.frame.color
Vector of length vcount(x).

vertex.rot
Vector of length vcount(x). Passed to polygons::npolygon, elevation degree from which the polygon is drawn.

vertex.frame.prop
Vector of length vcount(x). What proportion of the vertex does the frame occupy (values between 0 and 1).

edge.width
Vector of length ecount(x).

edge.width.range
Vector of length ecount(x).

edge.arrow.size
Vector of length ecount(x).

edge.color.mix
Numeric vector of length ecount(x) with values in [0,1]. 0 means color equal to ego's vertex color, one equals to alter's vertex color.

edge.color.alpha
Either a vector of length 1 or 2, or a matrix of size ecount(x)*2 with values in [0,1]. Alpha (transparency) levels (see details)

edge.curvature
Numeric vector of length ecount(x). Curvature of edges in terms of radians.

edge.line.lty
Vector of length ecount(x).

edge.line.breaks
Vector of length ecount(x). Number of vertices to draw (approximate) the arc (edge).

sample.edges
Numeric scalar between 0 and 1. Proportion of edges to sample.

skip.vertex
Logical scalar. When TRUE the object is not plotted.

skip.edges
Logical scalar. When TRUE the object is not plotted.

skip.arrows
Logical scalar. When TRUE the object is not plotted.

add
Logical scalar.

zero.margins
Logical scalar.